+++ packages/image/src/image-prop-types.js
@@ -1,8 +1,7 @@
 import PropTypes from "prop-types";
-import { Image } from "react-native";
 
 export const propTypes = {
-  ...Image.PropTypes,
+  style: PropTypes.object,
   accessibilityLabel: PropTypes.string,
   aspectRatio: PropTypes.number,
   borderRadius: PropTypes.number,
@@ -17,7 +16,6 @@ export const propTypes = {
 };
 
 export const defaultProps = {
-  ...Image.defaultImagePropTypes,
   aspectRatio: undefined,
   disablePlaceholder: false,
   fadeImageIn: false,