+++ packages/link/link.showcase.js
@@ -1,7 +1,7 @@
 import { TcView } from "@times-components/utils";
 import React from "react";
 import { colours } from "@times-components/styleguide";
-import Link, { TextLink, TimesTextLink } from "./src/link";
+import Link, { TextLink } from "./src/link";
 
 const BigSquare = props => (
   <TcView
@@ -97,92 +97,6 @@ export default {
         </TextLink>
       ),
       name: "TextLink with styles",
-      type: "story"
-    },
-    {
-      component: (_, { action }) => (
-        <Link onPress={action("onPress")} url="https://thetimes.co.uk">
-          <TcView
-            style={{
-              backgroundColor: colours.functional.backgroundSecondary,
-              height: 100,
-              width: 100
-            }}
-          />
-        </Link>
-      ),
-      name: "NEW: Link with big content",
-      type: "story"
-    },
-    {
-      component: (_, { action }) => {
-        const responsiveLinkStyles = {
-          base: `
-            height: 100px;
-            width: 100px;
-            display: block;
-            background-color: red;
-            color: white;
-          `,
-          medium: `
-            background-color: blue;
-          `
-        };
-
-        return (
-          <Link
-            onPress={action("onPress")}
-            url="https://thetimes.co.uk"
-            target="_blank"
-            responsiveLinkStyles={responsiveLinkStyles}
-            underlined={false}
-          >
-            resize me
-          </Link>
-        );
-      },
-      name: "NEW: Responsive Link",
-      type: "story",
-      platform: "web"
-    },
-    {
-      component: (_, { action }) => (
-        <Link
-          onPress={e => {
-            e.preventDefault();
-            action("onPress")(e);
-          }}
-          url="https://thetimes.co.uk"
-        >
-          <BigSquare />
-        </Link>
-      ),
-      name: "NEW: Link can prevent browser navigation",
-      type: "story"
-    },
-    {
-      component: (_, { action }) => (
-        <TimesTextLink
-          onPress={action("onPress")}
-          url="https://thetimes.co.uk/"
-        >
-          The Times
-        </TimesTextLink>
-      ),
-      name: "NEW: TextLink",
-      type: "story"
-    },
-    {
-      component: (_, { action }) => (
-        <TimesTextLink
-          onPress={action("onPress")}
-          style={{ color: colours.functional.action }}
-          url="https://thetimes.co.uk/"
-        >
-          The Times
-        </TimesTextLink>
-      ),
-      name: "NEW: TextLink with styles",
       type: "story"
     }
   ],