+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.78.2](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.78.1...@times-components/article-skeleton@1.78.2) (2022-04-08)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.78.1](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.78.0...@times-components/article-skeleton@1.78.1) (2022-04-08)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
 # [1.78.0](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.77.0...@times-components/article-skeleton@1.78.0) (2022-04-05)